move consumer document logics into a dedicated file
