+++ packages/agreement-process/src/services/agreementService.ts
@@ -734,60 +734,3 @@ export async function cloneAgreementLogic({
     events: [newAgreement, ...docEvents],
   };
 }
-
-export async function addConsumerDocumentLogic(
-  agreementId: string,
-  payload: ApiAgreementDocumentSeed,
-  agreementQuery: AgreementQuery,
-  authData: AuthData
-): Promise<CreateEvent<AgreementEvent>> {
-  const agreement = await agreementQuery.getAgreementById(agreementId);
-
-  assertAgreementExist(agreementId, agreement);
-  assertRequesterIsConsumer(agreement.data, authData);
-  assertCanWorkOnConsumerDocuments(agreement.data.state);
-
-  const existentDocument = agreement.data.consumerDocuments.find(
-    (d) => d.id === payload.id
-  );
-
-  if (existentDocument) {
-    throw agreementDocumentAlreadyExists(agreementId);
-  }
-
-  return toCreateEventAgreementConsumerDocumentAdded(
-    agreementId,
-    apiAgreementDocumentToAgreementDocument(payload),
-    agreement.metadata.version
-  );
-}
-
-export async function removeAgreementConsumerDocument(
-  agreementId: string,
-  documentId: string,
-  agreementQuery: AgreementQuery,
-  authData: AuthData,
-  fileRemove: (container: string, path: string) => Promise<void>
-): Promise<CreateEvent<AgreementEvent>> {
-  const agreement = await agreementQuery.getAgreementById(agreementId);
-
-  assertAgreementExist(agreementId, agreement);
-  assertRequesterIsConsumer(agreement.data, authData);
-  assertCanWorkOnConsumerDocuments(agreement.data.state);
-
-  const existentDocument = agreement.data.consumerDocuments.find(
-    (d) => d.id === documentId
-  );
-
-  if (!existentDocument) {
-    throw agreementDocumentNotFound(documentId, agreementId);
-  }
-
-  await fileRemove(config.storageContainer, existentDocument.path);
-
-  return toCreateEventAgreementConsumerDocumentRemoved(
-    agreementId,
-    documentId,
-    agreement.metadata.version
-  );
-}
